Use shared_ptr to a custom handler for read/write ops
Instead of using std::function. We can't use the latter because
CompletionTokens may be non-copyable operations (and std::function
requires whatever is passed to it to be copyable). And we wrap it into
shared_ptr because copies do actually need to be made inside Asio's
`post` function.

The handler can also serve to fix the problem with bound CompletionToken
executors which we currently don't utilize.