No rx_buffer resize
The assumption was that the resize would only change the internal size
number, but it seems it does the buffer reallocation and copy on each
such resize. Doing this speeds up the transmission by about 20% on slow
devices (e.g. phones). I'm guessing that the little time it spent after
a udp receive callback was executed and before another udp receive call
was scheduled has messed up the libutp's clocking which resulted in the
delays.

Might be worth moving the whole udp context work to another thread as
there may be other factors that mixup the libutp's timings.